Fix missing set-cookie issue.
It appears newer versions of FreeIPA (>4.8.6) do not return the set-cookie
header in every response. It's only included in the initial login response.
Here we were updating the sessionID on every rpc call and erroring out if the
set-cookie header was missing. As this is not necessary we simply return if the
header is missing.